Added all art I have done up till now, mostly linear
MOST of this is not the correct pixel size to be placed in unity, and
most the images are not transparent. The psd files should let you see
all layers - if the image looks messy/not finished, it’s because you’re
looking at all the layers, not just the final line art.